refactor: extract options in their own file
